# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Wasted API.",
        "title": "wasted",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Refresh access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Refresh code",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/request-code": {
            "post": {
                "description": "Send code to user email or phone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sent code user",
                "parameters": [
                    {
                        "description": "Login info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RequestCodeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.CodeSentResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify": {
            "post": {
                "description": "Verify code for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify code from user",
                "parameters": [
                    {
                        "description": "Login and code",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.VerifyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenPairResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.CodeSentData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Code sent"
                }
            }
        },
        "auth.CodeSentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.CodeSentData"
                }
            }
        },
        "auth.RefreshInput": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.RefreshResult"
                }
            }
        },
        "auth.RefreshResult": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "auth.RequestCodeInput": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string"
                }
            }
        },
        "auth.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.TokenPairResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/auth.TokenPair"
                }
            }
        },
        "auth.VerifyInput": {
            "type": "object",
            "required": [
                "code",
                "login"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                }
            }
        }
    }
}